Console logs to figure out the API for tiddlywiki.
isReady looks like it can just always return true.

getTiddlerInfo can return extra info for each tiddler if I wind
up needing it. Where this extra info goes is not super clear, but
possibly it goes in a field called 'bag' on the tiddler as by
default the 'bag' field is excluded from all saves in the filesystem
code. For now, it doesn't look like we'll need extra info, so
return {}; but maybe we'd like to save the docid and not have
the tiddler title be the docid.

getUpdatedTiddlers calls the callback to indicate new and modified
tiddlers when they occur on the server. Tiddlywiki doesn't assume
you'll be smart enough to call the callback multiple times, so each
call to the callback queues up another getUpdatedTiddlers call to
run one minute later. During the build process, this will cause
the build to loop forever if any updates come in, or at least to
loop until the callback stops getting called. For our purposes,
we can just listen to firebase and call the callback when there
are real changes, and make this fail/do nothing in node
mode for when the build is run so that the build completes.

saveTiddler is called each time a tiddler is saved. The standard
procedure for editing a tiddler XYZ is to create a new 'Draft of
XYZ' tiddler, and save it repeatedly while the user types. Then
when the draft is committed, XYZ is updated, and after that save
happens the draft is deleted. The draft's title is constant, and
the intended new title for the real tiddler is in a 'draft.title'
field.

loadTiddler is presumably called to get tiddler content for updated
tiddlers.

deleteTiddler is called to delete tiddlers, most often drafts after
the user is done editing, but sometimes an original tiddler if the
user clicks the trash icon.